Refactor: DataStore class definitions
