+++ src/datastores/index.ts
@@ -1,0 +1,8 @@
+//  ----------
+//  DataStores
+//  ----------
+
+export * from './base'
+export * from './fs'
+export * from './localStorage'
+export * from './memory'
