+++ src/index.ts
@@ -2,4 +2,6 @@
 //  DataStore
 //  ---------
 
-export * from './DataStore'
+export * from './adapters'
+export * from './transformers'
+export * from './datastores'
